==> 0_running_Jupyter_notebooks.ipynb <==
# Running a Jupyter notebook

This practical uses Jupyter notebooks, which are a great tool to organize, explain, and share your computational analyses and conclusions.  Jupyter notebooks can handle many coding languages, including R and python, both of which we will use today.  

### Let's spend a few minutes understanding how to navigate a Jupyter notebook.

## 1. Each one of these windows is a cell.  A cell can be one of two types:

1. _Markdown_ - used for typing text (m)
2. _Code_ - used for entering and running code (y)

Select this cell, press esc, and then press m (markdown) and then y (code).  The color coding changes, and in
code mode (y), you will see In [ ]: next to the window.  End in markdown mode (m), then press shift-return.


## 2. To run a cell, press shift-return

Try changing the cell types of these first windows and running them.  
If you try to run these cells in code (y) mode, you will get errors.

## 3. Cells can either be in editing or command mode

#### When you pressed **esc** above, you put the cell into command mode. 

In **command mode** (blue outlined cell), you can command the entire cell.  For example, you can move up and down between cells with the arrow keys and change the cell type with m and y (as we have seen).

#### If you press **enter** (or double click), you go into edit mode.

In **edit mode** (green outlined cell), you can edit cell contents.  

#### Try switching this cell between command and edit modes.

## 4. To get a new cell, press esc to get into command mode and then A or B, depending on whether you want a cell above or below the current one.

Try adding some blank cells.  To erase a cell, click on it, hit esc to get into command mode, and press dd

## 5. Now let's run some code.

Notebooks have a kernel - the underlying process that actually runs the code.  For this notebook, the kernel is R, but we will also use a python kernel later today.  If you get a message about the kernel crashing or needing to restart, don't panic.  Any variables you have already set will be lost, but you will be able to easily restarst the kernel and rerun the code to get them back.  If this happens, just let me know!

Run the following code cells by pressing shift-return. 

While a code cell is running, an asterisk will appear in the brackets to the left of the cell.  When it is done, a number will appear inside the brackets to denote the order in which the cell was run.

## This is a code cell

genotype <- seq(1,15,1)
print(genotype)

phenotype <- sample(1:20,15, replace=TRUE)
print(phenotype)

## What happens if we make a plot?

plot(genotype, phenotype)

## That's it!  Now we're ready to start with notebook 1_phenotype_exploration.ipynb.
